dlcoracle prepared for run within mit-dci/lit testing framework.
Added some command line parameters to the dlcoracle:
DataDir - so It could save keys in a test folder
HttpPort - to create multiple instances
Interval - to control the data publishing interval. To set the appropriate time when creating a r-point.

RangeFrom - Start interval of the publishing random variable
RangeTo - End interval of the publishing random variable